Release 1.4.16, 1.4.17, next
Added `next` to point to a not-yet-released version.
It's not listed in the dropdown, but navigable via an URL.

The point of having `next` folder is to be able to push things to
usermanual repo before the code lands in AT/AT main repo.

Also, a change to have both `1.x.x` and `latest` folders at the same time
in the repo, so that `1.x.x` is frozen, hence it's impossible to
inadvertently introduce some new development too early in `latest` folder
and making it later wrongly appear in `1.x.x` after next usermanual bump.

Any new development's documentation should be added to the `next` folder,
and then during the release time it'll be copied to `latest` accordingly.