feat ariatemplates/ariatemplates#1427 Customization of widget's error messages
Documentation for feature described there: ariatemplates/ariatemplates#1427

Not all details have been documented (widgets' resources are kind of fully private now).

Specific documentation for each widget has been added.

Below is the original description of the feature.

----

It is now possible for a user to customize the error messages used by the widgets.

There are 3 levels of error messages specifications, the two firsts being user configuration while the last is the hardcoded default.

This is resolved in this order of precedence:

- local: widget's configuration
- global: application's environment
- hardcoded: widget's resources